clean up teardown logic for demo and sandbox
