+++ cmd/demo/teardown.go
@@ -2,17 +2,12 @@ package demo
 
 import (
 	"context"
-	"fmt"
 
-	"github.com/flyteorg/flytectl/pkg/configutil"
+	"github.com/flyteorg/flytectl/pkg/sandbox"
 
 	"github.com/flyteorg/flytectl/pkg/docker"
 
-	"github.com/docker/docker/api/types"
-	"github.com/enescakir/emoji"
-
 	cmdCore "github.com/flyteorg/flytectl/cmd/core"
-	"github.com/flyteorg/flytectl/pkg/k8s"
 )
 
 const (
@@ -33,33 +28,5 @@ func teardownDemoCluster(ctx context.Context, args []string, cmdCtx cmdCore.Comm
 	if err != nil {
 		return err
 	}
-
-	return tearDownDemo(ctx, cli)
-}
-
-func tearDownDemo(ctx context.Context, cli docker.Docker) error {
-	c, err := docker.GetSandbox(ctx, cli)
-	if err != nil {
-		return err
-	}
-	if c != nil {
-		if err := cli.ContainerRemove(context.Background(), c.ID, types.ContainerRemoveOptions{
-			Force: true,
-		}); err != nil {
-			return err
-		}
-	}
-	if err := configutil.ConfigCleanup(); err != nil {
-		fmt.Printf("Config cleanup failed. Which Failed due to %v \n ", err)
-	}
-	if err := removeDemoKubeContext(); err != nil {
-		fmt.Printf("Kubecontext cleanup failed. Which Failed due to %v \n ", err)
-	}
-	fmt.Printf("%v %v Demo cluster is removed successfully. \n", emoji.Broom, emoji.Broom)
-	return nil
-}
-
-func removeDemoKubeContext() error {
-	k8sCtxMgr := k8s.NewK8sContextManager()
-	return k8sCtxMgr.RemoveContext(demoContextName)
+	return sandbox.Teardown(ctx, cli)
 }
